django/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "rdwatch"
version = "0.0.0dev0"
description = "A Django-based web application to support the Kitware WATCH project."
license = "Apache-2.0"
authors = [ "Kitware, Inc. <kitware@kitware.com>" ]
readme = "README.md"
homepage = "https://watch.resonantgeodata.com"
repository = "https://github.com/ResonantGeoData/RD-WATCH/django"
keywords = [ "django", "geospatial", "machine learning" ]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Web Environment",
  "Framework :: Django :: 4.0",
  "Framework :: Django",
]
packages = [ { include = "rdwatch", from = "src" } ]

[tool.poetry.dependencies]
python = "~3.10.0"
django = ">=4.1,<4.2"
djangorestframework = "~3.13.1"
psycopg2 = "~2.9.3"
redis = { version = "~4.3.4", extras = [ "hiredis" ] }
uritemplate = "^4.1.1"
iso3166 = "^2.1.1"
rio-tiler = "^3.1.6"
mercantile = "^1.2.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
black = "~22.6.0"
django-stubs = { version = "~1.12.0", extras = [ "compatible-mypy" ] }
djangorestframework-stubs = { version = "~1.7.0", extras = [ "compatible-mypy" ] }
isort = "~5.10.1"
mypy = "0.961"
taskipy = "^1.10.2"
flake8 = "^5.0.4"
flake8-django = "^1.1.5"


[tool.taskipy.tasks]
"build:openapi" = "DJANGO_SETTINGS_MODULE=rdwatch.server.settings django-admin generateschema --format openapi-json --file openapi.json"
"lint" = "task lint:black; task lint:isort; task lint:flake8"
"lint:fix" = "task lint:black:fix; task lint:isort:fix; task lint:flake8"
"lint:flake8" = "flake8 src"
"lint:black" = "black --check --diff --quiet src"
"lint:black:fix" = "black  --quiet src"
"lint:isort" = "isort --check --diff src"
"lint:isort:fix" = "isort src"
"lint:mypy" = "mypy src"


[tool.isort]
profile = "black"
known_first_party = ["rdwatch"]
known_django=["django", "rest_framework"]
sections=["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]


[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]


[[tool.mypy.overrides]]
module = [
  "django_filters",
  "mercantile"
]
ignore_missing_imports = true


[tool.django-stubs]
django_settings_module = "rdwatch.server.settings"


[tool.flake8]
extend_exclude = [".venv"]
extend_ignore = [
  # https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html?highlight=flake8#flake8
  "E203",
]
max_line_length = 88
